requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pillow==10.2.0
pydantic==2.5.3
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
